pyproject.toml
[tool.pdm]
[tool.pdm.dev-dependencies]
dev = [
    "alembic>=1.9.0",
    "setuptools>=65.6.3",
    "isort>=5.11.3",
    "wemake-python-styleguide>=0.17.0",
    "mypy>=0.991",
    "black[d]>=22.12.0",
    "pre-commit>=2.20.0",
    "types-passlib>=1.7.7.3",
    "types-python-jose>=3.3.4",
    "vulture>=2.6",
]

[project]
name = "expensa-backend"
version = "0.1.0"
description = "backend for expensa app"
authors = [{ name = "Yest Izit", email = "hi@dmitry.email" }]
dependencies = [
    "fastapi>=0.88.0",
    "uvicorn[standard]>=0.20.0",
    "pydantic>=1.10.2",
    "asyncpg>=0.27.0",
    "SQLAlchemy==2.0.0b4",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "greenlet>=2.0.1",
    "python-multipart>=0.0.5",
]
requires-python = ">=3.11"
license = { text = "MIT" }

[tool.setuptools]
py-modules = []

[tool.mypy]
exclude = [".git", ".idea", ".mypy_cache", ".venv", ".vscode", "migrations"]
pretty = true
check_untyped_defs = true
disallow_incomplete_defs = true
disallow_subclassing_any = true
disallow_untyped_calls = true
disallow_untyped_defs = true
warn_unreachable = true

[tool.isort]
profile = "black"

[tool.black]
exclude = [".git", ".idea", ".mypy_cache", ".venv", ".vscode", "migrations"]
line-length = 88

[tool.vulture]
exclude = [".git", ".idea", ".mypy_cache", ".venv", ".vscode", "migrations"]
ignore_names = ["Config", "frozen", "token_type", "orm_mode", "cls", "exp"]
ignore_decorators = ["@auth_router.*", "@user_router.*", "@validator"]
